Add support for metadata embedding in image files
Extending task setting output format: 'txt', 'json', or 'embed' If 'embed', the metadata will be embedded in PNG files in tEXt chunks, and as EXIF UserComment for JPEG files. Both can then be read back from the image using vanilla JavaScript.

This is the backend change. More updates on the way for the UI part, but this is a prerequisite. Also undoes PR 7 (Allow the UI to disable file saving), which I will add back later with a cleaner implementation.